requirements.txt
Django>=3.2
djangorestframework
gunicorn
openai
python-dotenv